feat(vm): single wazy engine, reflection-free host ABI
Replace wazero with wazy (github.com/samyfodil/wazy) as the only wasm
engine and remove the reflection the old multi-engine abstraction
required on the serving path.

- Engine: wazero -> wazy; wazero dropped from go.mod. wazy's typed host
  functions replace wazero's reflection-based WithFunc.
- Host ABI: core/vm gains generic HostFnN/HostProcN helpers; tools/hostfn-gen
  generates reflection-free HostFunctions() for the ~209 W_ methods across
  vm-low-orbit / vm-ops-orbit (make gen-hostfn), aggregating embedded
  providers. The satellite path passes wasm args as raw []uint64.
- Collapse: core/vm.{Module,Memory,Function,Global} alias wazy/api directly;
  delete the callBridge wrapper; drop the test-only reflection call path
  (Call(...interface{})/Return/Reflect), leaving FunctionInstance = RawCall.
- Layout: flatten pkg/vm/service/wazy + pkg/vm/helpers/wazy into pkg/vm
  (package vm); core/vm stays the interface package.

Warm Go->wasm call -25% time / -33% allocs; host-call adapter goes from
~1.2us/8-alloc to zero-alloc; full HTTP serving path warm allocs 207 -> 170.

+++ pkg/vm-orbit/tests/e2e/go/e2e_test.go
@@ -52,7 +52,7 @@ func TestFloat64Return(t *testing.T) {
 	assert.NilError(t, err)
 	defer plugin.Close()
 
-	ret := basicCall(t, plugin, path.Join(fixtureDir, "float_return.wasm"), uint32(5))
+	ret := basicCall(t, plugin, path.Join(fixtureDir, "float_return.wasm"), uint64(5))
 	testReturn(t, ret, 1)
 }
 
@@ -95,7 +95,7 @@ func TestUpdatePlugin(t *testing.T) {
 	module, err := testingSuite.WasmModule(basicWasm)
 	assert.NilError(t, err)
 
-	callVal := uint32(5)
+	callVal := uint64(5)
 	ret, err := module.Call(context.Background(), "ping", callVal)
 	assert.NilError(t, err)
